Rename column_aggregate_metric.py into column_aggregate_metric_provider.py for better code readability.
