+++ great_expectations/expectations/metrics/__init__.py
@@ -2,10 +2,10 @@ from .meta_metric_provider import (  # isort:skip
     MetaMetricProvider,
     DeprecatedMetaMetricProvider,
 )
-from .column_aggregate_metric import (
+from .column_aggregate_metric_provider import (
     ColumnMetricProvider,  # This class name is being deprecated (use "ColumnAggregateMetricProvider" going forward).
 )
-from .column_aggregate_metric import (
+from .column_aggregate_metric_provider import (
     ColumnAggregateMetricProvider,
     column_aggregate_partial,
     column_aggregate_value,
